+++ src/main/kotlin/com/tribus/markdown/actions/ExportHtmlAction.kt
@@ -4,7 +4,7 @@ import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.tribus.markdown.util.MarkdownFileUtil
 
-class ExportHtmlAction : AnAction() {
+class ExportHtmlAction : AnAction(), MarkdownAction {
     override fun actionPerformed(e: AnActionEvent) {
         // TODO: Phase 6 implementation
     }